refactor: Address code review suggestions
- Extract name generation to utils/nameGenerator.ts for better maintainability
- Add debouncing (300ms) for name replacement to fix race condition on rapid typing
- Add exponential backoff (500ms * attempt) between retry attempts
- Improve logging verbosity: redact raw LLM responses in production
- Use imported escapeRegex utility instead of inline function
